Move node code to nodelet file
