+++ zebROS_ws/src/tf_object_detection/CMakeLists.txt
@@ -13,6 +13,7 @@ find_package(catkin REQUIRED COMPONENTS
   field_obj_tracker
   image_transport
   message_filters
+  nodelet
   roscpp
   rospy
   sensor_msgs
@@ -142,7 +143,9 @@ include_directories(
 ## Declare a C++ executable
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
-add_executable(tf_object_screen_to_world src/screen_to_world_node.cpp src/depth_algorithms.cpp)
+
+## screen_to_world.cpp isn't a dependency, but I don't know how else to build it FIXME
+add_executable(tf_object_screen_to_world src/screen_to_world_node.cpp src/depth_algorithms.cpp src/screen_to_world.cpp)
 add_executable(test_depth_detection src/test_depth_detection.cpp src/depth_algorithms.cpp)
 add_executable(cargo_detection_filter src/cargo_filter.cpp)
 add_executable(dummy_camerainfo src/dummy_camerainfo.cpp)
